# frontend/Agricultural_Products_Market_admin/default_OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/cart": {
      "get": {
        "tags": [
          "购物车管理"
        ],
        "summary": "查看购物车列表",
        "description": "获取当前用户的购物车列表",
        "operationId": "getCartList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListShoppingCart"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "购物车管理"
        ],
        "summary": "修改购物车数量",
        "description": "更新购物车中商品的数量",
        "operationId": "updateCartQuantity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "购物车管理"
        ],
        "summary": "添加农产品到购物车",
        "description": "将农产品添加到当前用户的购物车",
        "operationId": "addToCart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/produce/{id}": {
      "put": {
        "tags": [
          "管理员农产品管理"
        ],
        "summary": "修改农产品",
        "description": "修改指定ID的农产品信息",
        "operationId": "updateProduce",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProduceRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "管理员农产品管理"
        ],
        "summary": "删除农产品",
        "description": "删除指定ID的农产品",
        "operationId": "deleteProduce",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/address/{id}": {
      "put": {
        "tags": [
          "地址管理"
        ],
        "summary": "修改收货地址",
        "description": "修改当前用户指定ID的收货地址",
        "operationId": "updateAddress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "地址管理"
        ],
        "summary": "删除收货地址",
        "description": "删除当前用户指定ID的收货地址",
        "operationId": "deleteAddress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/produce": {
      "get": {
        "tags": [
          "管理员农产品管理"
        ],
        "summary": "分页查询所有农产品",
        "description": "分页查询所有农产品，支持按名称模糊搜索",
        "operationId": "getProduceList_1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "produceName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultIPageProduceInfo"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "管理员农产品管理"
        ],
        "summary": "添加农产品",
        "description": "添加新的农产品信息",
        "operationId": "addProduce",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProduceRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/address": {
      "get": {
        "tags": [
          "地址管理"
        ],
        "summary": "查看地址列表",
        "description": "获取当前用户的所有收货地址",
        "operationId": "getAddressList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListUserAddress"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "地址管理"
        ],
        "summary": "添加收货地址",
        "description": "为当前用户添加新的收货地址",
        "operationId": "addAddress",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAddressDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/produce": {
      "get": {
        "tags": [
          "农产品管理"
        ],
        "summary": "获取农产品列表",
        "description": "获取上架农产品列表，支持按名称模糊查询",
        "operationId": "getProduceList",
        "parameters": [
          {
            "name": "produceName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "currentPage",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageDTOListProduceInfo"
                }
              }
            }
          }
        }
      }
    },
    "/produce/{id}": {
      "get": {
        "tags": [
          "农产品管理"
        ],
        "summary": "获取农产品详情",
        "description": "获取指定ID的农产品详细信息",
        "operationId": "getProduceDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProduceInfo"
                }
              }
            }
          }
        }
      }
    },
...
    "/cart/{id}": {
      "delete": {
        "tags": [
          "购物车管理"
        ],
        "summary": "删除购物车项",
        "description": "删除购物车中指定ID的商品",
        "operationId": "deleteCartItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "type": "string"
          }
        }
      },
      "UpdateCartDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ProduceRequestDTO": {
        "type": "object",
        "properties": {
          "produceName": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64"
          },
          "manufacturer": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "coverUrl": {
            "type": "string"
          }
        }
      },
...
      "UpdateAddressDTO": {
        "type": "object",
        "properties": {
          "recipient": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "isDefault": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "AddCartDTO": {
        "type": "object",
        "properties": {
          "produceId": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "AddAddressDTO": {
        "type": "object",
        "properties": {
          "recipient": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "isDefault": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ResultListShoppingCart": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShoppingCart"
            }
          }
        }
      },
      "ShoppingCart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "购物车项ID",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "description": "用户ID",
            "format": "int64"
          },
          "produceId": {
            "type": "integer",
            "description": "农产品ID",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "description": "农产品数量",
            "format": "int32"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "更新时间",
            "format": "date-time"
          }
        },
        "description": "购物车表"
      },
      "ProduceInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "农产品ID",
            "format": "int64"
          },
          "produceName": {
            "type": "string",
            "description": "农产品名称"
          },
          "categoryId": {
            "type": "integer",
            "description": "分类ID",
            "format": "int64"
          },
          "manufacturer": {
            "type": "string",
            "description": "生产商"
          },
          "price": {
            "type": "number",
            "description": "农产品价格"
          },
          "stock": {
            "type": "integer",
            "description": "库存数量",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "description": "农产品简介"
          },
          "status": {
            "type": "integer",
            "description": "状态：1-上架，0-下架",
            "format": "int32"
          },
          "coverUrl": {
            "type": "string",
            "description": "封面图片URL"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "更新时间",
            "format": "date-time"
          }
        },
        "description": "农产品信息表"
      },
      "PageDTOListProduceInfo": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "总条数",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "description": "总页码数",
            "format": "int64"
          },
          "list": {
            "type": "array",
            "description": "当前页数据",
            "items": {
              "type": "array",
              "description": "当前页数据",
              "items": {
                "$ref": "#/components/schemas/ProduceInfo"
              }
            }
          }
        },
        "description": "分页结果"
      },
      "ResultPageDTOListProduceInfo": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageDTOListProduceInfo"
          }
        }
      },
      "ResultProduceInfo": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ProduceInfo"
          }
        }
      },
...
      "IPageProduceInfo": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProduceInfo"
            }
          },
          "current": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ResultIPageProduceInfo": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/IPageProduceInfo"
          }
        }
      },
      "ResultListUserAddress": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserAddress"
            }
          }
        }
      },
      "UserAddress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "地址ID",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "description": "用户ID",
            "format": "int64"
          },
          "recipient": {
            "type": "string",
            "description": "收件人姓名"
          },
          "phone": {
            "type": "string",
            "description": "联系电话"
          },
          "address": {
            "type": "string",
            "description": "详细地址"
          },
          "isDefault": {
            "type": "integer",
            "description": "是否默认地址：1-是，0-否",
            "format": "int32"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          }
        },
        "description": "用户地址"
      }
    }
  },
...
}